examples/dtls-echo: fix debug message

+++ examples/dtls-echo/dtls-client.c
@@ -161,7 +161,7 @@ static int _peer_get_psk_info_handler(struct dtls_context_t *ctx,
     switch (type) {
         case DTLS_PSK_IDENTITY:
             if (id_len) {
-                dtls_debug("got psk_identity_hint: '%.*s'\n", id_len, id);
+                dtls_debug("got psk_identity_hint: '%.*s'\n", (int)id_len, id);
             }
 
             if (result_length < psk_id_length) {
